sql: make setup of leaf txn for streamer more bullet-proof
This commit makes sure that we try to use the streamer API only if we
can actually create a non-nil `LeafTxn`. In some edge cases it appears
that all the previous checks were insufficient, and this commit should
take care of that. Note that I couldn't reproduce those edge cases
manually, so there is no regression test.

Release note: None
